backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary>=2.9.10
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.27.2
phonenumbers==8.13.36
email-validator==2.2.0
redis==5.2.0
qrcode[pil]==8.0
gtts==2.5.1
pyttsx3==2.90

